Update package for tracing package 1.0-beta1

+++ Sources/TracingOpenTelemetrySemanticConventions/SpanAttributes+CodeSemantics.swift
@@ -64,6 +64,6 @@ public struct CodeAttributes: SpanAttributeNamespace {
         /// E.g. 42.
         ///
         /// You may want to set this attribute to `#line`.
-        public var lineNumber: Key<Int> { "code.lineno" }
+        public var lineNumber: Key<Int64> { "code.lineno" }
     }
 }
